[wip] Implementing #tick for DualSmelterBlockEntity
Followed the rough outline of AbstractFurnaceBlockEntity, since this will do
the same type of cooking.  However, this needs a recipe with multiple inputs,
so it will need to hybridize the idea of fuel and burning with the multi-input
of a ShapedRecipe.  Not yet clear whether I want to reuse that or not.