Set cooking times correctly
AbstractFurnace sets these as part of manipulating itself due to implementing
Inventory.  Since Dual Smelter doesn't implement Inventory, it has-a Inventory,
another approach is necessary.  Therefore, initialize cook time and set total
cook time when the cooking starts, not prior to cooking starting.  AbstractFurnace
sets the total cooking time when the input is provided, which works, but results
in setting that time in multiple functions due to needing to reset it if a second
item cooks immediately following the first one.