Show Rose Gold Ingot in the RecipeBook
The ClientRecipeBook and RecipeBookWidget are tightly
controlled via private and private static methods.  They
are not easy to extend, so use Mixin to pry open the ClientRecipeBook
and allow returning an appropriate RecipeBookGroup enum.

If I added many more multiple-input smelting recipes, it might make
sense to look at EarlyRiser to hack the enums that construct the
RecipeBookWidget.  This would allow either a separate tab for Furnace
or a new category altogether.

Next steps: 1) textures, 2) multi-count input/output, 3) XP drop, and
4) figure out why the recipe isn't unlocking once crafted.